restructure files
- move login modal to homepage
- rename chinguApplication to register
- extract formCreator style to its own style sheet
