+++ src/components/Register/Register.css
@@ -52,7 +52,6 @@ ORIGINAL PROJECT VARIABLES
   background-color: rgba(0, 0, 0, 0.164); }
 
 .chingu-application-modal {
-  margin-top: 100px;
   display: block;
   z-index: 110;
   width: 350px;
@@ -60,4 +59,6 @@ ORIGINAL PROJECT VARIABLES
   border-radius: 20px;
   box-shadow: 0px 3px 6px #E6E6E6;
   margin: 0 auto;
-  padding: 10px 20px; }
+  margin-top: 100px;
+  margin-bottom: 100px;
+  padding: 20px; }
