+++ src/components/VoyageApplication/VoyageApplication.css
@@ -54,253 +54,3 @@ ORIGINAL PROJECT VARIABLES
   padding: 20px;
   margin-top: 100px;
   box-shadow: 0px 3px 6px #E6E6E6; }
-
-/*
-==========================
-QUESTION & ANSWER UNIT
-==========================
-*/
-.voyage-application-QA {
-  margin-top: 20px; }
-
-/*
-==========================
-QUESTION
-==========================
-*/
-.voyage-application-question {
-  font-size: 14px;
-  color: #757575; }
-
-/*
-==========================
-QUESTION SUBTEXT
-==========================
-*/
-.voyage-application-subtext {
-  font-size: 12px;
-  color: #BBBBBB; }
-
-/*
-==========================
-ANSWER FONT
-==========================
-*/
-.voyage-application-answer {
-  font-size: 14px;
-  color: #757575; }
-
-/*
-==========================
-INPUT & TEXTAREA
-==========================
-*/
-.voyage-application-input {
-  display: block;
-  width: 100%;
-  font-size: 14px;
-  color: #757575;
-  padding: 5px;
-  margin-top: 5px;
-  border: 1px solid #E6E6E6; }
-
-.voyage-application-text-area {
-  display: block;
-  width: 100%;
-  font-size: 14px;
-  color: #757575;
-  padding: 5px;
-  margin-top: 5px;
-  border: 1px solid #E6E6E6;
-  min-height: 50px;
-  max-height: 50px;
-  max-width: 100%;
-  min-width: 100%; }
-
-/*
-==========================
-DROPDOWN
-==========================
-*/
-.voyage-application-dropdown {
-  width: 100%;
-  border: 1px solid #E6E6E6;
-  padding: 5px;
-  color: #757575;
-  font-size: 12px;
-  margin-top: 5px; }
-
-.voyage-application-dropdown > option {
-  width: 418px !important; }
-
-/*
-==========================
-DROPDOWN w/ CHECKBOX
-==========================
-*/
-.search-input-box {
-  margin-bottom: 10px;
-  min-width: 180px;
-  border: none;
-  box-shadow: 0px 3px 6px #E6E6E6;
-  padding: 5px; }
-
-.search-input-box::placeholder {
-  color: #E6E6E6; }
-
-.filter-dropdown-box {
-  background-color: white;
-  border: 1px solid #E6E6E6;
-  width: 180px;
-  display: none;
-  position: absolute;
-  box-shadow: 0px 3px 6px #E6E6E6;
-  z-index: 10;
-  max-height: 120px;
-  overflow: scroll;
-  overflow-x: hidden;
-  text-align: left;
-  color: #757575;
-  font-size: 12px;
-  margin-left: -9px; }
-
-.filter-dropdown-hide {
-  display: none; }
-
-.filter-dropdown-btn-5,
-.filter-dropdown-btn-101 {
-  min-width: 180px;
-  color: #757575;
-  font-size: 12px;
-  background-color: white;
-  border: 1px solid #E6E6E6;
-  padding: 7px; }
-
-.filter-dropdown-btn-5:hover #dropdown_multiple-5 {
-  display: block; }
-
-.filter-dropdown-btn-101:hover #dropdown_multiple-101 {
-  display: block; }
-
-/*
-==========================
-CHECKBOX STYLING
-==========================
-*/
-.checkbox-container {
-  display: block;
-  position: relative;
-  padding-left: 35px;
-  margin-bottom: 12px;
-  cursor: pointer;
-  -webkit-user-select: none;
-  -moz-user-select: none;
-  -ms-user-select: none;
-  user-select: none; }
-
-/* Hide the browser's default checkbox */
-.checkbox-container input {
-  position: absolute;
-  opacity: 0;
-  cursor: pointer; }
-
-/* Create a custom checkbox */
-.checkmark {
-  position: absolute;
-  margin-left: 10px;
-  top: 0;
-  left: 0;
-  height: 14px;
-  width: 14px;
-  background-color: white;
-  border: 1px solid #15CF89; }
-
-/* On mouse-over, add a grey background color */
-.checkbox-container:hover input ~ .checkmark {
-  background-color: #15CF89; }
-
-/* When the checkbox is checked, add a blue background */
-.checkbox-container input:checked ~ .checkmark {
-  background-color: #15CF89; }
-
-/* Create the checkmark/indicator (hidden when not checked) */
-.checkmark:after {
-  content: '';
-  position: absolute;
-  display: none; }
-
-/* Show the checkmark when checked */
-.checkbox-container input:checked ~ .checkmark:after {
-  display: block; }
-
-/* Style the checkmark/indicator */
-.checkbox-container .checkmark:after {
-  left: 4px;
-  top: 1.5px;
-  width: 4px;
-  height: 7px;
-  border: solid white;
-  border-width: 0 2px 2px 0;
-  -webkit-transform: rotate(45deg);
-  -ms-transform: rotate(45deg);
-  transform: rotate(45deg); }
-
-/*
-==========================
-RADIO STYLING
-==========================
-*/
-.radio-container {
-  display: block;
-  position: relative;
-  padding-left: 35px;
-  margin-bottom: 12px;
-  cursor: pointer;
-  -webkit-user-select: none;
-  -moz-user-select: none;
-  -ms-user-select: none;
-  user-select: none; }
-
-/* Hide the browser's default radio button */
-.radio-container input {
-  position: absolute;
-  opacity: 0; }
-
-/* Create a custom radio button */
-.radio-checkmark {
-  position: absolute;
-  margin-left: 11px;
-  top: 0;
-  left: 0;
-  height: 14px;
-  width: 14px;
-  background-color: white;
-  border-radius: 50%;
-  border: 1px solid #15CF89; }
-
-/* On mouse-over, add a grey background color */
-.radio-container:hover input ~ .radio-checkmark {
-  background-color: #15CF89; }
-
-/* When the radio button is checked, add a blue background */
-.radio-container input:checked ~ .radio-checkmark {
-  background-color: #15CF89; }
-
-/* Create the indicator (the dot/circle - hidden when not checked) */
-.radio-checkmark:after {
-  content: '';
-  position: absolute;
-  display: none; }
-
-/* Show the indicator (dot/circle) when checked */
-.radio-container input:checked ~ .radio-checkmark:after {
-  display: block; }
-
-/* Style the indicator (dot/circle) */
-.radio-container .radio-checkmark:after {
-  top: 2.5px;
-  left: 2.5px;
-  width: 7px;
-  height: 7px;
-  border-radius: 50%;
-  background: white; }
